python_requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.7
action-tutorials-py==0.20.7
actionlib-msgs==4.9.0
Adafruit-Blinka==8.68.0
adafruit-circuitpython-ads1x15==3.0.2
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-icm20x==2.1.8
adafruit-circuitpython-mcp230xx==2.5.19
adafruit-circuitpython-motor==3.4.18
adafruit-circuitpython-mpu6050==1.3.5
adafruit-circuitpython-pca9685==3.4.20
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-servokit==1.3.22
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.85.0
Adafruit-PureIO==1.1.11
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
appdirs==1.4.4
argcomplete==1.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
binho-host-adapter==0.1.6
blinker==1.4
bloom==0.13.0
breezy==3.2.1
Brotli==1.0.9
builtin-interfaces==1.2.2
catkin-pkg==1.1.0
catkin-pkg-modules==1.1.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==25.2
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.20.1
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.2
configobj==5.0.6
constantly==15.1.0
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cv-bridge==3.2.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
demo-nodes-py==0.20.7
Deprecated==1.2.13
diagnostic-msgs==4.9.0
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
fastbencode==0.0.5
fastimport==0.9.14
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
geometry-msgs==4.9.0
gpg==1.16.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
interactive-markers==2.3.2
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
lark==1.1.1
laser-geometry==2.4.1
launch==1.0.13
launch-ros==0.19.13
launch-testing==1.0.13
launch-testing-ros==0.19.13
launch-xml==1.0.13
launch-yaml==1.0.13
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.2
logging-demo==0.20.7
lxml==4.8.0
lz4==3.1.3+dfsg
map-msgs==2.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mercurial==6.1.1
meson==1.10.0
message-filters==4.3.11
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
nav-msgs==4.9.0
netifaces==0.11.0
nmea-msgs==2.0.0
nmea-navsat-driver==2.0.1
nose2==0.9.2
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
osrf-pycommon==2.1.6
packaging==21.3
patiencediff==0.2.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.7
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.8.0
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
pyftdi==0.57.1
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyrsistent==0.18.1
pyserial==3.5
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gitlab==2.10.1
python-magic==0.4.24
python-qt-binding==1.1.3
pythran==0.10.0
pytz==2022.1
pyusb==1.3.1
PyYAML==5.4.1
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.7
rcl-interfaces==1.2.2
rclpy==3.3.19
rcutils==5.1.7
requests==2.25.1
requests-toolbelt==0.9.1
resource-retriever==3.1.3
rmw-dds-common==1.6.0
roman==3.3
ros2action==0.18.15
ros2bag==0.15.15
ros2cli==0.18.15
ros2component==0.18.15
ros2doctor==0.18.15
ros2interface==0.18.15
ros2launch==0.19.13
ros2lifecycle==0.18.15
ros2multicast==0.18.15
ros2node==0.18.15
ros2param==0.18.15
ros2pkg==0.18.15
ros2run==0.18.15
ros2service==0.18.15
ros2topic==0.18.15
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosdep==0.26.0
rosdep-modules==0.26.0
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rospkg-modules==1.6.0
RPi.GPIO==0.7.1
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scipy==1.8.0
SecretStorage==3.3.1
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
service-identity==18.1.0
shape-msgs==4.9.0
six==1.16.0
smbus2==0.5.0
snowballstemmer==2.2.0
sos==4.9.2
soupsieve==2.3.1
sros2==0.10.7
ssh-import-id==5.11
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.9
systemd-python==234
sysv-ipc==1.1.0
teleop-twist-keyboard==2.4.1
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.17
tf2-kdl==0.25.17
tf2-msgs==0.25.17
tf2-py==0.25.17
tf2-ros-py==0.25.17
tf2-sensor-msgs==0.25.17
tf2-tools==0.25.17
toml==0.10.2
topic-monitor==0.20.7
trajectory-msgs==4.9.0
transforms3d==0.3.1
turtlesim==1.4.3
Twisted==22.1.0
typing_extensions==4.15.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
vcstool==0.3.0
vcstools==0.1.42
visualization-msgs==4.9.0
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.13.3
zipp==1.0.0
zope.interface==5.4.0
